+++ lib/govspeak/html_sanitizer.rb
@@ -55,7 +55,7 @@ class Govspeak::HtmlSanitizer
       elements: Sanitize::Config::RELAXED[:elements] + %w[govspeak-embed-attachment govspeak-embed-attachment-link svg path].concat(allowed_elements),
       attributes: {
         :all => Sanitize::Config::RELAXED[:attributes][:all] + %w[role aria-label],
-        "a" => Sanitize::Config::RELAXED[:attributes]["a"] + [:data],
+        "a" => Sanitize::Config::RELAXED[:attributes]["a"] + [:data] + %w[draggable],
         "svg" => Sanitize::Config::RELAXED[:attributes][:all] + %w[xmlns width height viewbox focusable],
         "path" => Sanitize::Config::RELAXED[:attributes][:all] + %w[fill d],
         "div" => [:data],